More refactor
I don't actually know what I'm doing.
